refactor: camelCase in urls
BREAKING CHANGE: old urls with hyphen delimiter are not working anymore
