+++ src/services/users/__docs__/confirmEmail.yaml
@@ -1,4 +1,4 @@
-/users/confirm-email:
+/users/confirmEmail:
   get:
     tags:
       - 'users'
