+++ src/services/users/__docs__/refreshToken.yaml
@@ -1,4 +1,4 @@
-/users/refresh-token:
+/users/refreshToken:
   post:
     tags:
       - 'users'
