+++ src/services/users/__test__/signin.test.ts
@@ -17,7 +17,7 @@ describe('POST /users/signin', () => {
     const user = await User.findOne({ where: { id: response.body.user.id } })
     if (user != null) {
       await request(app)
-        .get(`/users/confirm-email?tempToken=${user.tempToken as string}`)
+        .get(`/users/confirmEmail?tempToken=${user.tempToken as string}`)
         .send()
         .expect(200)
     }
