+++ src/services/users/index.ts
@@ -4,7 +4,7 @@ import { confirmEmailRouter } from './confirmEmail'
 import { currentRouter } from './current'
 import { OAuth2Router } from './oauth2'
 import { refreshTokenRouter } from './refreshToken'
-import { resetPasswordRouter } from './reset-password'
+import { resetPasswordRouter } from './resetPassword'
 import { signinRouter } from './signin'
 import { signoutRouter } from './signout'
 import { signupRouter } from './signup'
