+++ src/services/users/resetPassword/__docs__/post.yaml
@@ -1,4 +1,4 @@
-/users/reset-password:
+/users/resetPassword:
   post:
     tags:
       - 'users'
