+++ src/services/users/resetPassword/__docs__/put.yaml
@@ -1,9 +1,9 @@
-/users/reset-password:
+/users/resetPassword:
   put:
     tags:
       - 'users'
     summary: 'Change the user password'
-    description: 'Change the user password if the provided tempToken (sent in the email of POST /users/reset-password) is correct.'
+    description: 'Change the user password if the provided tempToken (sent in the email of POST /users/resetPassword) is correct.'
     requestBody:
       content:
         application/json:
